requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic[email]==2.9.2
python-dateutil==2.9.0
aiosmtplib==3.0.2
jinja2==3.1.4
python-dotenv==1.0.1
